NF: QueueType becomes an enum
This improve the typing.

+++ AnkiDroid/src/main/java/com/ichi2/anki/AnkiDroidJsAPI.kt
@@ -251,7 +251,7 @@ open class AnkiDroidJsAPI(private val activity: AbstractFlashcardViewer) {
             "cardLeft" -> convertToByteArray(apiContract, currentCard.left)
             "cardODid" -> convertToByteArray(apiContract, currentCard.oDid)
             "cardODue" -> convertToByteArray(apiContract, currentCard.oDue)
-            "cardQueue" -> convertToByteArray(apiContract, currentCard.queue)
+            "cardQueue" -> convertToByteArray(apiContract, currentCard.queue.toInt())
             "cardLapses" -> convertToByteArray(apiContract, currentCard.lapses)
             "cardDue" -> convertToByteArray(apiContract, currentCard.due)
             "deckName" -> convertToByteArray(apiContract, Decks.basename(activity.getColUnsafe.decks.name(currentCard.did)))
